uqmi: add fixed timeouts to calls to uqmi
Koen Vanderputte writes:

===
Now the bug is that all modems using qmi do not reply
on the 1st "sync" but they do on subsequent.
"No reply" as in : qmi does not send anything back on
the bus the 1ste time.

There are 2 fixes needed in uqmi to fix this once and for all

- send "sync" twice
- use "poll" to wait for 2 FD's
  - a pipe FD which can be set by a signal handler
  - the fd of the qmi device

Set the timeout to 20 seconds for:

- scan command

Set the timeout to 5 seconds for:

- connect command (WDS)
- Retry this command 5x more when receiving "call failed"
  (when still searching for a network, they tend to send this error)

Set the timeout to 1 sec for all others.

I'm using this way of working in my code using this generated lib and
it has been working for years without any hang or hiccup on hundreds
of devices from 3G up to 5G SA and on various modems (quectel, sierra, Telit, tdtech, fibocom, ..)
===

Adapt qmi netifd script according to this writeup, by search-and-replace
with following Vim expression:

:%s/uqmi -s -d/uqmi -s -t 1 -d/g

And then manually set the timeouts to:
- 20s for "--network-register" command
- 5s for "--start-network" command.

Now, a retry scheme shall be implemented in subsequent commit for the
"start network" command - the registration already does have one - and
the sleeps can probably be removed while at that.

Signed-off-by: Lech Perczak <lech.perczak@gmail.com>